requirements.txt
flask
flask-cors
flask-bcrypt
pyjwt
